boot/i686-bios: Add basic structure and build system
The shell scripts for building currently only do the minimum work that
they have to do; yet there is a help command.

The README.md explains the project  structure and where to find all of
the code. The .gitignore  of boot/i686-bios consists  mostly of binary
data that would otherwise clutter the repository.