tools/machinize: Identify type of statement in parser
The parser can now identify what kind  of statement is presented to it
at a specific point in  time and can convert the CLI (Clear Interrupt)
flag to the  corresponding machine code  byte. Furthermore, the parser
can understand and store labels to look them up later.

The code is still highly imperfect, the error handling questionable in
utility and the functionality lacking everywhere, but it works.